requirements.txt
requests
python-dotenv
fal-client
httpx
httpcore
Pillow
numpy
torch
